+++ connection_test.go
@@ -496,6 +496,9 @@ func TestServerClientCancellation(t *testing.T) {
 	opts.DefaultConnectionOptions.SendCancelOnContextCanceled = true
 	opts.DefaultConnectionOptions.PropagateCancel = true
 
+	serverStats := newRecordingStatsReporter()
+	opts.StatsReporter = serverStats
+
 	testutils.WithTestServer(t, opts, func(t testing.TB, ts *testutils.TestServer) {
 		callReceived := make(chan struct{})
 		testutils.RegisterFunc(ts.Server(), "ctxWait", func(ctx context.Context, args *raw.Args) (*raw.Res, error) {
@@ -519,16 +522,23 @@ func TestServerClientCancellation(t *testing.T) {
 		_, _, _, err := raw.Call(ctx, ts.Server(), ts.HostPort(), ts.ServiceName(), "ctxWait", nil, nil)
 		assert.Equal(t, ErrRequestCancelled, err, "client call result")
 
+		statsTags := ts.Server().StatsTags()
+		serverStats.Expected.IncCounter("inbound.cancels.requested", statsTags, 1)
+		serverStats.Expected.IncCounter("inbound.cancels.honored", statsTags, 1)
+
 		calls := relaytest.NewMockStats()
 		calls.Add(ts.ServiceName(), ts.ServiceName(), "ctxWait").Failed("canceled").End()
 		ts.AssertRelayStats(calls)
 	})
+
+	serverStats.ValidateExpected(t)
 }
 
 func TestCancelWithoutSendCancelOnContextCanceled(t *testing.T) {
 	tests := []struct {
 		msg                         string
 		sendCancelOnContextCanceled bool
+		wantCancelRequested         bool
 	}{
 		{
 			msg:                         "no send or process cancel",
@@ -537,6 +547,7 @@ func TestCancelWithoutSendCancelOnContextCanceled(t *testing.T) {
 		{
 			msg:                         "only enable cancels on outbounds",
 			sendCancelOnContextCanceled: true,
+			wantCancelRequested:         true,
 		},
 	}
 
@@ -545,7 +556,12 @@ func TestCancelWithoutSendCancelOnContextCanceled(t *testing.T) {
 			opts := testutils.NewOpts()
 			opts.DefaultConnectionOptions.SendCancelOnContextCanceled = tt.sendCancelOnContextCanceled
 
+			serverStats := newRecordingStatsReporter()
+			opts.StatsReporter = serverStats
+
 			testutils.WithTestServer(t, opts, func(t testing.TB, ts *testutils.TestServer) {
+				serverStats.Reset()
+
 				callReceived := make(chan struct{})
 				testutils.RegisterFunc(ts.Server(), "ctxWait", func(ctx context.Context, args *raw.Args) (*raw.Res, error) {
 					require.NoError(t, ctx.Err(), "context valid before cancellation")
@@ -571,6 +587,17 @@ func TestCancelWithoutSendCancelOnContextCanceled(t *testing.T) {
 				calls := relaytest.NewMockStats()
 				calls.Add(ts.ServiceName(), ts.ServiceName(), "ctxWait").Failed("timeout").End()
 				ts.AssertRelayStats(calls)
+
+				ts.AddPostFn(func() {
+					// Validating these at the end of the test, when server has fully processed the cancellation.
+					if tt.wantCancelRequested && !ts.HasRelay() {
+						serverStats.Expected.IncCounter("inbound.cancels.requested", ts.Server().StatsTags(), 1)
+						serverStats.ValidateExpected(t)
+					} else {
+						serverStats.EnsureNotPresent(t, "inbound.cancels.requested")
+					}
+					serverStats.EnsureNotPresent(t, "inbound.cancels.honored")
+				})
 			})
 		})
 	}